@@ -1,7 +1,6 @@
 package services
 
 import (
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"runtime"
@@ -41,7 +40,7 @@ func (c *loginService) OpenBrowser(url string) error {
 
 func (c *loginService) WriteToConfigFile(configPath string, data string) error {
 	// Write file as 0600 since it contains private keys.
-	return ioutil.WriteFile(configPath, []byte(data), 0600)
+	return os.WriteFile(configPath, []byte(data), 0600)
 }
 
 func (c *loginService) DeleteConfigFile(configPath string) error {
