@@ -1,16 +1,31 @@
 package app
 
-import "github.com/urfave/cli/v2"
+import (
+	"fmt"
+	"runtime/debug"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/urfave/cli/v2"
+)
 
 const (
-	DefaultVersion = "v0.13.0"
+	pseudoVersionMinLen        = len("vX.0.0-yyyymmddhhmmss-abcdefabcdef")
+	pseudoVersionCommitInfoLen = len("yyyymmddhhmmss-abcdefabcdef")
 )
 
 var (
-	BuildDate string
-	Commit    string
-	Version   string
+	date    string
+	commit  string
+	version string
 )
+
+type BuildInfo struct {
+	Date    string // build time or commit time.
+	Commit  string
+	Version string
+}
 
 func NewVersionCommand() (CommandOut, error) {
 	return CommandOut{Command: &cli.Command{
@@ -18,23 +33,64 @@ func NewVersionCommand() (CommandOut, error) {
 		Usage:   "Version information",
 		Aliases: []string{"v"},
 		Action: func(c *cli.Context) error {
-			return PrintObj(&struct {
-				BuildDate string
-				Commit    string
-				Version   string
-			}{
-				BuildDate: BuildDate,
-				Commit:    Commit,
-				Version:   getVersion(),
-			})
+			return PrintObj(NewBuildInfo())
 		},
 	}}, nil
 }
 
-func getVersion() string {
-	version := Version
-	if len(version) == 0 {
-		version = DefaultVersion
+// NewBuildInfo will populate build info, to make debugging API errors easier,
+// in the three scenarios a user can install tcld:
+//  1. Installed via the makefile or via brew.
+//  2. Installed via `go install`.
+//  3. Compiled on a branch via `go build ./cmd/tcld`
+func NewBuildInfo() BuildInfo {
+	if len(version) > 0 {
+		// Used when built with make or installed with brew.
+		return BuildInfo{
+			Date:    date,
+			Commit:  commit,
+			Version: version,
+		}
 	}
-	return version
+
+	di, ok := debug.ReadBuildInfo()
+	if !ok {
+		fmt.Printf("Failed to read debug info\n")
+		return BuildInfo{}
+	}
+
+	info := BuildInfo{
+		Version: di.Main.Version,
+	}
+	if len(di.Main.Version) >= pseudoVersionMinLen {
+		// Used when compiled with `go install`.
+		// See https://go.dev/ref/mod#pseudo-versions for more info on the expected string format
+		commitInfoStart := len(di.Main.Version) - pseudoVersionCommitInfoLen
+		split := strings.Split(di.Main.Version[commitInfoStart:], "-")
+
+		// Make the time human readable.
+		at, err := time.Parse("20060102150405", split[0])
+		if err == nil {
+			info.Date = at.UTC().Format("2006-01-02T15:04:05.000Z")
+		}
+		info.Commit = split[1]
+	} else {
+		// Used when built directly from a branch with `go build`.
+		var hash, modified string
+		for _, setting := range di.Settings {
+			switch setting.Key {
+			case "vcs.revision":
+				hash = setting.Value
+			case "vcs.modified":
+				if v, err := strconv.ParseBool(setting.Value); err == nil && v {
+					modified = "-modified"
+				}
+			case "vcs.time":
+				info.Date = setting.Value
+			}
+		}
+		info.Commit = fmt.Sprintf("%s%s", hash, modified)
+	}
+
+	return info
 }
